Ticket #287: density and internal energy stopping conditions support, implemented for Fi + tests
git-svn-id: file:///home/amuse/svn/amuse/trunk@3365 6e832d89-e7d1-4ad1-8ca1-873beecccf65

+++ lib/stopcond/stopcond.c
@@ -293,14 +293,14 @@ int set_stopping_condition_minimum_density_parameter(double value) {
     minimum_density_parameter = value;
     return 0;
 }
-int set_stopping_condition_minimum_density_parameter_(int *value) {
+int set_stopping_condition_minimum_density_parameter_(double *value) {
     return set_stopping_condition_minimum_density_parameter(*value);
 }
 int get_stopping_condition_minimum_density_parameter(double *value) {
     *value = minimum_density_parameter;
     return 0;
 }
-int get_stopping_condition_minimum_density_parameter_(int *value) {
+int get_stopping_condition_minimum_density_parameter_(double *value) {
     return get_stopping_condition_minimum_density_parameter(value);
 }
 
@@ -312,14 +312,14 @@ int set_stopping_condition_maximum_density_parameter(double value) {
     }
     return 0;
 }
-int set_stopping_condition_maximum_density_parameter_(int *value) {
+int set_stopping_condition_maximum_density_parameter_(double *value) {
     return set_stopping_condition_maximum_density_parameter(*value);
 }
 int get_stopping_condition_maximum_density_parameter(double *value) {
     *value = maximum_density_parameter;
     return 0;
 }
-int get_stopping_condition_maximum_density_parameter_(int *value) {
+int get_stopping_condition_maximum_density_parameter_(double *value) {
     return get_stopping_condition_maximum_density_parameter(value);
 }
 
@@ -327,14 +327,14 @@ int set_stopping_condition_minimum_internal_energy_parameter(double value) {
     minimum_internal_energy_parameter = value;
     return 0;
 }
-int set_stopping_condition_minimum_internal_energy_parameter_(int *value) {
+int set_stopping_condition_minimum_internal_energy_parameter_(double *value) {
     return set_stopping_condition_minimum_internal_energy_parameter(*value);
 }
 int get_stopping_condition_minimum_internal_energy_parameter(double *value) {
     *value = minimum_internal_energy_parameter;
     return 0;
 }
-int get_stopping_condition_minimum_internal_energy_parameter_(int *value) {
+int get_stopping_condition_minimum_internal_energy_parameter_(double *value) {
     return get_stopping_condition_minimum_internal_energy_parameter(value);
 }
 
@@ -346,14 +346,14 @@ int set_stopping_condition_maximum_internal_energy_parameter(double value) {
     }
     return 0;
 }
-int set_stopping_condition_maximum_internal_energy_parameter_(int *value) {
+int set_stopping_condition_maximum_internal_energy_parameter_(double *value) {
     return set_stopping_condition_maximum_internal_energy_parameter(*value);
 }
 int get_stopping_condition_maximum_internal_energy_parameter(double *value) {
     *value = maximum_internal_energy_parameter;
     return 0;
 }
-int get_stopping_condition_maximum_internal_energy_parameter_(int *value) {
+int get_stopping_condition_maximum_internal_energy_parameter_(double *value) {
     return get_stopping_condition_maximum_internal_energy_parameter(value);
 }
